requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.8
django-crispy-forms==2.1
django-js-asset==2.2.0
django-mptt==0.15.0
django-seed==0.3.1
environs==10.0.0
Faker==21.0.0
marshmallow==3.20.1
packaging==23.2
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2023.3
